infrastructure/requirements.txt
paramiko
fabric
ansible
docker
flask
requests
pyyaml
